Add authorization Add healthcheck Remove ClamAV scan
